Enable GPU-to-GPU comm in TensorPipeAgent
This commit uses TensorPipe's cuda_ipc channel to conduct
cross-process same-machine GPU-to-GPU communication. On the sender
side, `TensorPipeAgent` grabs a stream to each device, let these
streams wait for current streams, and passes the streams to
TensorPipe `CudaBuffer`. On the receiver side, it also grabs a stream
for each device, and uses these streams to receive tensors and run
user functions. After that, these streams are then used for sending
the response back to the sender. When receiving the response, the
sender will grab a new set of streams and use them for TensorPipe's
`CudaBuffer`.

If device maps are provided, `TensorPipeAgent::send` will return a
derived class of `CUDAFuture`, which is specifically tailored for
RPC Messages.

TODOs:
1. When sending a request, RPC caller only need to grab streams for
used tensor devices in the message, which is different from the callee.
2. Enable sending CUDA RPC to the same process.
3. Add a custom CUDA stream pool.

ghstack-source-id: ab4f5567a38c9a12a6a57c39c2c23b05f4ae9bc1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44418